Refactor to make test suite leaner
