+++ tests/unit_tests/gurobipy_var_dicts_funcs/var_funcs_test.py
@@ -9,23 +9,11 @@ from gurobipy import GRB
 
 from opti_extensions import IndexSet1D, IndexSetND, ParamDict1D, ParamDictND
 from opti_extensions.gurobipy import addVars
+from tests.unit_tests._shared_var_funcs_tests import *  # noqa: F401, F403
+from tests.unit_tests._shared_var_funcs_tests import SHARED_INDEXSET_CASES
 
 
-@pytest.mark.parametrize('indexset', [IndexSet1D(), IndexSetND()])
-def test_addVars_indexset_valerr(mdl, indexset):
-    with pytest.raises(ValueError):
-        addVars(mdl, indexset, vtype=GRB.CONTINUOUS)
-
-
-indexset_cases = [
-    IndexSet1D(['A', 'B', 'C']),
-    IndexSet1D(range(3)),
-    IndexSetND(['A', 'B', 'C'], range(3)),
-    IndexSetND(range(3), range(3)),
-]
-
-
-@pytest.mark.parametrize('indexset', indexset_cases)
+@pytest.mark.parametrize('indexset', SHARED_INDEXSET_CASES)
 @pytest.mark.parametrize(
     'vtype', [GRB.CONTINUOUS, GRB.BINARY, GRB.INTEGER, GRB.SEMICONT, GRB.SEMIINT]
 )
@@ -36,33 +24,6 @@ def test_addVars_pass(mdl, mdl_2, indexset, vtype):
     mdl_2.update()
 
     assert repr(dict(one)) == repr(two)
-
-
-@pytest.mark.parametrize('mdl', ['abc', 123, ('A', 'B')])
-@pytest.mark.parametrize('indexset', indexset_cases)
-@pytest.mark.parametrize(
-    'vtype', [GRB.CONTINUOUS, GRB.BINARY, GRB.INTEGER, GRB.SEMICONT, GRB.SEMIINT]
-)
-def test_addVars_mdl_typerr(mdl, indexset, vtype):
-    with pytest.raises(TypeError):
-        addVars(mdl, indexset, vtype=vtype)
-
-
-@pytest.mark.parametrize(
-    'indexset',
-    [
-        (1, 2, 3),
-        [1, 2, 3],
-        [('A', 'B'), ('C', 'D')],
-        [('A', '1'), ('C', '2')],
-    ],
-)
-@pytest.mark.parametrize(
-    'vtype', [GRB.CONTINUOUS, GRB.BINARY, GRB.INTEGER, GRB.SEMICONT, GRB.SEMIINT]
-)
-def test_addVars_idxset_typerr(mdl, indexset, vtype):
-    with pytest.raises(TypeError):
-        addVars(mdl, indexset, vtype=vtype)
 
 
 @pytest.mark.parametrize(
@@ -82,31 +43,3 @@ def test_addVars_paramdict_bound(mdl, indexset, paramdict, bound_type):
 
     mdl.update()
     assert repr(dict(one)) == repr(two)
-
-
-@pytest.mark.parametrize(
-    'indexset, paramdict',
-    [
-        (IndexSet1D(['A', 'B', 'C']), ParamDictND({('A', 0): 1, ('B', 0): 1, ('C', 0): 1})),
-        (IndexSetND(range(2), range(2)), ParamDict1D({0: 1, 1: 1})),
-    ],
-)
-@pytest.mark.parametrize('bound_type', ['lb', 'ub'])
-def test_addVars_paramdict_typerr(mdl, indexset, paramdict, bound_type):
-    with pytest.raises(TypeError):
-        bound_kwargs_1 = {bound_type: paramdict}
-        addVars(mdl, indexset, vtype=GRB.CONTINUOUS, **bound_kwargs_1)
-
-
-@pytest.mark.parametrize(
-    'indexset, paramdict',
-    [
-        (IndexSetND(range(2), range(2)), ParamDictND({(0, 0, 0): 1, (0, 1, 0): 1})),
-        (IndexSetND(range(2), range(2)), ParamDictND({(0, 0, 0, 0): 1, (0, 1, 0, 0): 1})),
-    ],
-)
-@pytest.mark.parametrize('bound_type', ['lb', 'ub'])
-def test_addVars_paramdict_difflen_valerr(mdl, indexset, paramdict, bound_type):
-    with pytest.raises(ValueError):
-        bound_kwargs_1 = {bound_type: paramdict}
-        addVars(mdl, indexset, vtype=GRB.CONTINUOUS, **bound_kwargs_1)
