+++ tests/unit_tests/highspy_var_dicts_funcs/vardictND_subset_selection_test.py
@@ -4,116 +4,4 @@
 
 """Subset selection methods of VarDictND."""
 
-import pytest
-from highspy import HighsVarType
-
-from opti_extensions.highspy import addVariables
-
-
-@pytest.mark.parametrize(
-    '_input, values, expected',
-    [
-        ('setNd_cmb2', ('*', 1), [(0, 1), (1, 1)]),
-        ('setNd_cmb2', (0, '*'), [(0, 0), (0, 1)]),
-        ('setNd_cmb2', (2, '*'), []),
-        ('setNd_cmb2', ('*', 2), []),
-        ('setNd_cmb3', (0, '*', '*'), [(0, 0, 0), (0, 0, 1), (0, 1, 0), (0, 1, 1)]),
-        ('setNd_cmb3', ('*', 1, '*'), [(0, 1, 0), (0, 1, 1), (1, 1, 0), (1, 1, 1)]),
-        ('setNd_cmb3', (0, 1, '*'), [(0, 1, 0), (0, 1, 1)]),
-        ('setNd_cmb3', (0, '*', 1), [(0, 0, 1), (0, 1, 1)]),
-        ('setNd_cmb3', (2, '*', '*'), []),
-        ('setNd_cmb3', (0, '*', 2), []),
-    ],
-)
-def test_subset_keys_pass(mdl, request, _input, values, expected):
-    indexset = request.getfixturevalue(_input)
-    v = addVariables(mdl, indexset, type=HighsVarType.kContinuous, name_prefix='VAL')
-
-    assert v.subset_keys(*values) == expected
-
-
-@pytest.mark.parametrize('values', [(0, '*'), (0, '*', 2, '*')])
-def test_subset_keys_diff_tuplelen(mdl, setNd_cmb3, values):
-    v = addVariables(mdl, setNd_cmb3, type=HighsVarType.kContinuous, name_prefix='VAL')
-    with pytest.raises(ValueError):
-        v.subset_keys(*values)
-
-
-@pytest.mark.parametrize('values', [((0, 1), (0, 0)), [(0, 1)]])
-def test_subset_keys_nonscaler(mdl, setNd_cmb2, values):
-    v = addVariables(mdl, setNd_cmb2, type=HighsVarType.kContinuous, name_prefix='VAL')
-    with pytest.raises(TypeError):
-        v.subset_keys(*values)
-
-
-@pytest.mark.parametrize('values', [('*', '*'), (0, 1)])
-def test_subset_keys_invalid_values(mdl, setNd_cmb2, values):
-    v = addVariables(mdl, setNd_cmb2, type=HighsVarType.kContinuous, name_prefix='VAL')
-    with pytest.raises(ValueError):
-        v.subset_keys(*values)
-
-
-@pytest.mark.parametrize(
-    '_input, values, expected',
-    [
-        ('setNd_cmb2', ('*', 1), [1, 3]),
-        ('setNd_cmb2', (0, '*'), [0, 1]),
-        ('setNd_cmb2', (2, '*'), []),
-        ('setNd_cmb2', ('*', 2), []),
-        ('setNd_cmb3', (0, '*', '*'), [0, 1, 2, 3]),
-        ('setNd_cmb3', ('*', 1, '*'), [2, 3, 6, 7]),
-        ('setNd_cmb3', (0, 1, '*'), [2, 3]),
-        ('setNd_cmb3', (0, '*', 1), [1, 3]),
-        ('setNd_cmb3', (2, '*', '*'), []),
-        ('setNd_cmb3', (0, '*', 2), []),
-    ],
-)
-def test_subset_values_pass(mdl, request, _input, values, expected):
-    indexset = request.getfixturevalue(_input)
-    v = addVariables(mdl, indexset, type=HighsVarType.kContinuous, name_prefix='VAL')
-
-    # In Highs, variables are added sequentially.
-    # For setNd_cmb2 (2x2), indices are 0,1,2,3.
-    # (0,0)->0, (0,1)->1, (1,0)->2, (1,1)->3.
-    # For setNd_cmb3 (2x2x2), indices are 0..7.
-    # (0,0,0)->0, (0,0,1)->1, (0,1,0)->2, (0,1,1)->3, (1,0,0)->4, (1,0,1)->5, (1,1,0)->6, (1,1,1)->7
-
-    # The expected values in the original test seem to be indices?
-    # In Gurobi test, expected values were [1, 3] etc.
-    # But wait, addVars returns Var objects.
-    # The original test compares
-    # `v.subset_values(*values) == [v[x] for x in v.subset_keys(*values)]`.
-    # It doesn't check against `expected` list of integers directly.
-    # Ah, the `expected` argument in `test_subset_values_pass` is NOT used in the assertion!
-    # It is just passed but ignored?
-    # Let's check the Gurobi test again.
-    # `assert v.subset_values(*values) == [v[x] for x in v.subset_keys(*values)]`
-    # Yes, `expected` is unused in the body. It was probably copy-pasted from
-    # `test_subset_keys_pass` and left there.
-    # I will remove `expected` from parameters and parametrization to clean it up, or just leave
-    # it unused to match structure.
-    # I'll leave it unused to keep it simple and consistent with the source I'm adapting, but I'll
-    # fix the assertion logic to be correct for Highs.
-
-    assert v.subset_values(*values) == [v[x] for x in v.subset_keys(*values)]
-
-
-@pytest.mark.parametrize('values', [(0, '*'), (0, '*', 2, '*')])
-def test_subset_values_diff_tuplelen(mdl, setNd_cmb3, values):
-    v = addVariables(mdl, setNd_cmb3, type=HighsVarType.kContinuous, name_prefix='VAL')
-    with pytest.raises(ValueError):
-        v.subset_values(*values)
-
-
-@pytest.mark.parametrize('values', [((0, 1), (0, 0)), [(0, 1)]])
-def test_subset_values_nonscaler(mdl, setNd_cmb2, values):
-    v = addVariables(mdl, setNd_cmb2, type=HighsVarType.kContinuous, name_prefix='VAL')
-    with pytest.raises(TypeError):
-        v.subset_values(*values)
-
-
-@pytest.mark.parametrize('values', [('*', '*'), (0, 1)])
-def test_subset_values_invalid_values(mdl, setNd_cmb2, values):
-    v = addVariables(mdl, setNd_cmb2, type=HighsVarType.kContinuous, name_prefix='VAL')
-    with pytest.raises(ValueError):
-        v.subset_values(*values)
+from tests.unit_tests._shared_vardictND_subset_selection_tests import *  # noqa: F401, F403
